// src/types/routes.ts
/** @format */

export enum RoutesPath {
  LOGIN = "/",
  REGISTER = "/singup",
  PROFILES = "/profiles",
  CREATEPROFILES = "/profiles/create",
  EDITPROFILE = "/profiles/edit/:id",
  HOMEPAGE = "/profile/homepage/:id",
  GAMESLIST = "/profile/homepage/list/:id",
  MYGAMES = "/profile/homepage/library",
  CREATEGAMES = "/profile/createGames&Genrer/:id",
  CREATEGENRES = "/profile/genrers/:id",
  GAMEOPT = "/profile/game/:id",
  EDITGAME = "/profile/game/edit/:id",
}